Add menu functionality Add anchoring Add Screen size scaling Add Input Field Add Buttons

+++ Assets/Scripts/Plate/PlateMode.cs
@@ -5,13 +5,24 @@ using System.Linq;
 using System.Threading.Tasks;
 using Unity.Jobs;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class PlateMode : Mode
 {
     public static List<Position> draw_hover;
     public static List<Position> draw_selected;
     protected static bool waiting;
-    List<Menu> m_menu;
+    protected static bool short_cuts_off;
+    EditPlateMenu m_menu;
+    protected EditPlateMenu GetPlateMenu(Plate selected_plate)
+    {
+        if (m_menu == null)
+        {
+            m_menu = Menu.CreateMenu<EditPlateMenu>();
+        }
+        m_menu.SetPlate(selected_plate);
+        return m_menu;
+    }
     void Start()
     {
         if (current_menu == null)
@@ -19,54 +30,61 @@ public class PlateMode : Mode
             SetUp();
         }
         Initialise();
-        m_menu = new List<Menu>();
-        for(int i = 0; i < 1; i++)
-        {
-            m_menu.Add(Menu.CreateMenu<EditPlateMenu>());
-            m_menu[i].SetOffset(192*i, 0);
-        }
+    }
+
+    public void Initialise()
+    {
+        Mode.current_menu = this;
+        selected = new List<Clickable>();
+        draw_hover = new List<Position>();
+        draw_selected = new List<Position>();
+        //m_menu = new List<Menu>();
+        Tile.Clicked += OnClick;
+        Tile.Hovered += OnHover;
+        EditPlateMenu.Clicked += OnClick;
+        TextField.Selected += Focus;
+        TextField.UnSelected += Unfocus;
+        SaveMenu.Save += OnClick;
+        SaveMenu save = Menu.CreateMenu<SaveMenu>();
     }
 
     void Update()
     {
-        MoveCamera();
-        if(Input.GetKeyDown( KeyCode.Space))
+        if (!short_cuts_off)
         {
-            SaveFile<Plate>(Plate.plate.Values.ToList());
+            MoveCamera();
         }
-        if(Input.GetKeyDown(KeyCode.Return))
-        {
-            List<Plate> load = LoadFile<Plate>();
-            Plate.plate.Clear();
-            for(int i = 0; i < load.Count; i++)
-            {
-                Plate.CreatePlate((load[i]));
-            }
-        }
-        if(!Input.GetKeyDown(KeyCode.Mouse0) && Input.GetKeyDown(KeyCode.Mouse1))
-        {
+        if (!Input.GetKeyDown(KeyCode.Mouse0) && Input.GetKeyDown(KeyCode.Mouse1))
+        { 
             if (selected.Count > 0)
             {
-                selected[selected.Count - 1].UnSelect();
-                selected[selected.Count - 1].UnHover();
-                List<Position> unselected = draw_selected;
-                if (selected.Count > 1)
+                if (selected[0] is Tile)
                 {
-                    int index = draw_selected.IndexOf(((Tile)selected[selected.Count - 2]).position);
-                    unselected = draw_selected.GetRange(index + 1, draw_selected.Count - index - 1);
-                    draw_selected.RemoveRange(index + 1, draw_selected.Count - index - 1);
+                    selected[selected.Count - 1].UnSelect();
+                    selected[selected.Count - 1].UnHover();
+                    List<Position> unselected = draw_selected;
+                    if (selected.Count > 1)
+                    {
+                        int index = draw_selected.IndexOf(((Tile)selected[selected.Count - 2]).position);
+                        unselected = draw_selected.GetRange(index + 1, draw_selected.Count - index - 1);
+                        draw_selected.RemoveRange(index + 1, draw_selected.Count - index - 1);
+                    }
+                    else
+                    {
+                        draw_selected.Clear();
+                        ClearHover();
+                    }
+                    for (int i = 0; i < unselected.Count; i++)
+                    {
+                        Map.map[unselected[i]].UnSelect();
+                    }
+                    selected.RemoveAt(selected.Count - 1);
                 }
-                else
+                else if (selected[0] is Plate)
                 {
-                    draw_selected.Clear();
-                    ClearHover();
+                    m_menu.UnSelect();
+                    Clear();
                 }
-                for (int i = 0; i < unselected.Count; i++)
-                {
-                    Map.map[unselected[i]].UnSelect();
-                }
-                selected.RemoveAt(selected.Count - 1);
-                OnHover();
             }
         }
     }
@@ -77,7 +95,7 @@ public class PlateMode : Mode
 
         public int GetDistance(Position from, Position to)
         {
-            if (Map.map[to].GetContinent() != null)
+            if (Map.map[to].GetContinent() != Guid.Empty)
             {
                 return -1;
             }
@@ -136,22 +154,39 @@ public class PlateMode : Mode
         }
     }
 
-    public override void OnSelect()
+    public void OnClick(Tile given_tile)
     {
-        if (selected.Count > 1)
+        selected.Add(given_tile);
+        if (given_tile.GetContinent() != Guid.Empty)
         {
-            Tile destination = (Tile)selected[selected.Count - 1];
-            if (draw_hover.Count > 0 && draw_hover[draw_hover.Count - 1] == destination.position)
+            try
             {
-                if (draw_selected.Contains(destination.position))
+                Clear();
+                Plate selected_plate = Plate.plate[given_tile.GetContinent()];
+                GetPlateMenu(selected_plate);
+                selected.Add(selected_plate);
+                selected[0].OnSelect();
+                m_menu.OnSelect();
+                Debug.Log(selected[0] + " " + selected_plate.id);
+            }
+            catch(Exception e)
+            {
+                Debug.Log(given_tile + " " + e);
+            }
+        }
+        else if (selected.Count > 1)
+        {
+            if (draw_hover.Count > 0 && draw_hover[draw_hover.Count - 1] == given_tile.position)
+            {
+                if (draw_selected.Contains(given_tile.position))
                 {
                     draw_selected.AddRange(draw_hover);
-                    draw_hover.Clear();
+                    ClearHover();
                     try
                     {
                         Plate.CreatePlate(new List<Position>(draw_selected));
                     }
-                    catch(Exception e)
+                    catch (Exception e)
                     {
                         Plate.CreatePlate(new List<Position>(draw_selected));
                         throw e;
@@ -174,33 +209,52 @@ public class PlateMode : Mode
                 selected.RemoveAt(selected.Count - 1);
             }
         }
-        else if(selected[0] is Tile)
+
+    }
+
+    public void OnClick(EditPlateMenu given_menu)
+    {
+        Debug.Log("Menu event started");
+    }
+
+    public void OnClick(bool save)
+    {
+        if (save)
         {
-            Tile current_tile = selected[0] as Tile;
-            if (current_tile.GetContinent() != null)
+            SaveFile<Plate>(Plate.plate.Values.ToList());
+        }
+        else
+        {
+            List<Plate> load = LoadFile<Plate>();
+            Plate.plate.Clear();
+            for (int i = 0; i < load.Count; i++)
             {
-                Clear();
-                selected.Add(Plate.plate[current_tile.GetContinent()]);
-                selected[0].OnSelect();
-                Debug.Log(selected[0]);
+                Plate.CreatePlate((load[i]));
             }
         }
     }
 
-    public override void OnHover()
+
+    public void OnHover(Tile given_tile)
     {
         if (selected.Count > 0)
         {
-            if (selected[selected.Count-1] is Tile && hover is Tile)
+            if (selected[selected.Count-1] is Tile )
             {
                 Tile start = (Tile) selected[selected.Count - 1];
-                Tile destination = (Tile)hover;
+                Tile destination = given_tile;
                 List<Position> draw_copy = new List<Position>(draw_hover);
                 PlateDrawContoller.EndFindPath();
                 ClearHover();
                 PlateDrawContoller.ScheduleFindPath(start, destination);
             }
         }
+        given_tile.OnHover();
+    }
+
+    public void UnHover(Tile given_tile)
+    {
+        given_tile.UnHover();
     }
 
     public static void ClearHover()
@@ -217,18 +271,19 @@ public class PlateMode : Mode
         for (int i = 0; i < draw_selected.Count; i++)
         {
             Map.map[draw_selected[i]].UnSelect();
-            Map.map[draw_selected[i]].UnHover();
         }
         draw_selected.Clear();
         base.Clear();
     }
 
-    public void Initialise()
+    public void Focus(TextField given_field)
     {
-        Mode.current_menu = this;
-        selected = new List<Clickable>();
-        draw_hover = new List<Position>();
-        draw_selected = new List<Position>();
+        short_cuts_off = true;
+    }
+
+    public void Unfocus(TextField given_field)
+    {
+        short_cuts_off = false;
     }
 
     public override void OnModeEnter()
